Move helper to different package
