+++ setup.cfg
@@ -20,6 +20,7 @@ install_requires =
 
 [options.packages.find]
 include = inmanta_plugins*
+    inmanta_files
 
 [flake8]
 # H101 Include your name with TODOs as in # TODO(yourname). This makes it easier to find out who the author of the comment was.
